+++ broker/src/functional-test/groovy/com/swisscom/cloud/sb/broker/functional/ServiceBrokerMetricsServiceFunctionalSpec.groovy
@@ -1,5 +1,10 @@
 package com.swisscom.cloud.sb.broker.functional
 
+import com.swisscom.cloud.sb.broker.metrics.BindingMetricsService
+import com.swisscom.cloud.sb.broker.metrics.LastOperationMetricsService
+import com.swisscom.cloud.sb.broker.metrics.LifecycleTimeMetricsService
+import com.swisscom.cloud.sb.broker.metrics.ServiceBrokerMetricsConfig
+import com.swisscom.cloud.sb.broker.metrics.ServiceInstanceMetricsService
 import com.swisscom.cloud.sb.broker.model.repository.CFServiceRepository
 import com.swisscom.cloud.sb.broker.model.repository.PlanRepository
 import com.swisscom.cloud.sb.broker.model.repository.ServiceBindingRepository
@@ -8,6 +13,7 @@ import com.swisscom.cloud.sb.broker.util.Resource
 import com.swisscom.cloud.sb.broker.util.measurements.*
 import com.swisscom.cloud.sb.client.ServiceBrokerClientExtended
 import com.swisscom.cloud.sb.client.model.DeleteServiceInstanceRequest
+import groovy.util.logging.Slf4j
 import org.influxdb.InfluxDB
 import org.influxdb.InfluxDBFactory
 import org.influxdb.dto.Query
@@ -16,13 +22,11 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.cloud.servicebroker.model.CreateServiceInstanceBindingRequest
 import org.springframework.cloud.servicebroker.model.CreateServiceInstanceRequest
 import org.springframework.web.client.RestTemplate
-import uk.org.lidalia.slf4jext.LoggerFactory
 
 import static junit.framework.Assert.assertEquals
 
 class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
 
-    final String DB_NAME = "mydb"
     final String SERVICE_NAME = "service1Name"
     final String SERVICE_GUID = "service1GuidIntegrationTest"
     final String PLAN_NAME = "small"
@@ -37,6 +41,8 @@ class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
     InfluxDBResultMapper influxDBResultMapper
 
     @Autowired
+    ServiceBrokerMetricsConfig serviceBrokerMetricsConfig
+    @Autowired
     CFServiceRepository cfServiceRepository
     @Autowired
     ServiceInstanceRepository serviceInstanceRepository
@@ -46,161 +52,93 @@ class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
     PlanRepository planRepository
 
     def setup() {
-        serviceBrokerClientExtended = new ServiceBrokerClientExtended(new RestTemplate(), "http://localhost:8080", serviceLifeCycler.cfAdminUser.username, serviceLifeCycler.cfAdminUser.password, serviceLifeCycler.cfExtUser.username, serviceLifeCycler.cfExtUser.password)
+        serviceBrokerClientExtended = new ServiceBrokerClientExtended(
+                new RestTemplate(),
+                "http://localhost:8080",
+                serviceLifeCycler.cfAdminUser.username,
+                serviceLifeCycler.cfAdminUser.password,
+                serviceLifeCycler.cfExtUser.username,
+                serviceLifeCycler.cfExtUser.password)
 
-        influxDB = InfluxDBFactory.connect("http://localhost:8086", "admin", "admin")
-        if (influxDB.databaseExists(DB_NAME)) {
-            influxDB.deleteDatabase(DB_NAME)
+        serviceBrokerClientExtended.createOrUpdateServiceDefinition(Resource.readTestFileContent("/service-data/service1.json"))
+        Thread.sleep(WAIT_FOR_SERVICE_DEFINITION_TO_BE_DONE_IN_MILLISECONDS)
+
+        influxDB = InfluxDBFactory.connect(
+                serviceBrokerMetricsConfig.uri,
+                serviceBrokerMetricsConfig.userName,
+                serviceBrokerMetricsConfig.password)
+
+        if (!influxDB.databaseExists(serviceBrokerMetricsConfig.db)) {
+           influxDB.createDatabase(serviceBrokerMetricsConfig.db)
         }
-        influxDB.createDatabase(DB_NAME)
 
         influxDBResultMapper = new InfluxDBResultMapper()
     }
 
     def cleanup() {
-        influxDB.deleteDatabase(DB_NAME)
         influxDB.close()
     }
 
     def "only measurements that are retrievable from the database are recorded when db is empty and initialized with 0"() {
         given:
         def metricsRetrievableFromDB = [
-                new MeasurementPoint("binding_total_total", BindingTotalTotalPoint),
-                new MeasurementPoint("provisionRequest_fail_ratio", ProvisionRequestFailRatioPoint),
-                new MeasurementPoint("provisionRequest_success_ratio", ProvisionRequestSuccessRatioPoint),
-                new MeasurementPoint("provisionRequest_total_fail", ProvisionRequestTotalFailPoint),
-                new MeasurementPoint("provisionRequest_total_success", ProvisionRequestTotalSuccessPoint),
-                new MeasurementPoint("provisionRequest_total_total", ProvisionRequestTotalTotalPoint),
-                new MeasurementPoint("provisionedInstances_fail_ratio", ProvisionedInstancesFailRatioPoint),
-                new MeasurementPoint("provisionedInstances_success_ratio", ProvisionedInstancesSuccessRatioPoint),
-                new MeasurementPoint("provisionedInstances_total_fail", ProvisionedInstancesTotalFailPoint),
-                new MeasurementPoint("provisionedInstances_total_success", ProvisionedInstancesTotalSuccessPoint),
-                new MeasurementPoint("provisionedInstances_total_total", ProvisionedInstancesTotalTotalPoint)
-        ]
-
-        and:
-        def dynamicallyGeneratedMetrics = [
-                new MeasurementPoint("bindingRequest_service_fail_${SERVICE_NAME}", BindingRequestServiceFailServicePoint),
-                new MeasurementPoint("bindingRequest_service_success_${SERVICE_NAME}", BindingRequestServiceSuccessServicePoint),
-                new MeasurementPoint("bindingRequest_service_total_${SERVICE_NAME}", BindingRequestServiceTotalServicePoint),
-                new MeasurementPoint("binding_service_total_${SERVICE_NAME}", BindingServiceTotalServicePoint),
-                new MeasurementPoint("lifecycleTime_service_total_${SERVICE_NAME}", LifecycleTimeServiceTotalServicePoint)
+                new MeasurementPoint(name: BindingMetricsService.BINDING_SERVICE_KEY),
+                new MeasurementPoint(name: BindingMetricsService.NEW_SERVICE_BINDINGS_KEY),
+                new MeasurementPoint(name: LastOperationMetricsService.LAST_OPERATION_EVENTS_KEY),
+                new MeasurementPoint(name: LastOperationMetricsService.LAST_OPERATIONS_KEY),
+                new MeasurementPoint(name: LifecycleTimeMetricsService.LIFECYCLE_TIME_KEY),
+                new MeasurementPoint(name: ServiceInstanceMetricsService.SERVICE_INSTANCES_KEY),
         ]
 
         and:
         Thread.sleep(TIME_TO_WAIT_FOR_WRITING_TO_INFLUXDB_TO_OCCUR_IN_MILLISECONDS)
 
         when:
-        Query query = new Query("show measurements", DB_NAME)
+        Query query = new Query("show measurements", serviceBrokerMetricsConfig.db)
         def queryResult = influxDB.query(query)
 
         and:
         def influxDBResultMapper = new InfluxDBResultMapper()
-        def results = influxDBResultMapper.toPOJO(queryResult, MeasurementPoint.class)
-
-        and:
-        def measurementValueList = new ArrayList()
-        results.name.each { measurement ->
-            query = new Query("select value from ${measurement}", "mydb")
-            def valueQueryResult = influxDB.query(query)
-            measurementValueList.add(valueQueryResult.getResults().get(0).getSeries().get(0).getValues().get(0).get(1))
-        }
+        List<MeasurementPoint> results = influxDBResultMapper.toPOJO(queryResult, MeasurementPoint.class)
 
         then:
-        assert (results.size() == metricsRetrievableFromDB.size())
+        assert (results.size() >= metricsRetrievableFromDB.size())
         assert (results.name.containsAll(metricsRetrievableFromDB.name))
-        assert (!results.name.containsAll(dynamicallyGeneratedMetrics.name))
-
-        and:
-        assert (measurementValueList.each {
-            result ->
-                result == 0.0
-        })
-    }
-
-    def "write dynamically generated metrics upon service definition and initialize with zero"() {
-        given:
-        def metrics = [
-                new MeasurementPoint("bindingRequest_service_fail_${SERVICE_NAME}", BindingRequestServiceFailServicePoint, 0),
-                new MeasurementPoint("bindingRequest_service_success_${SERVICE_NAME}", BindingRequestServiceSuccessServicePoint, 0),
-                new MeasurementPoint("bindingRequest_service_total_${SERVICE_NAME}", BindingRequestServiceTotalServicePoint, 0),
-                new MeasurementPoint("binding_service_total_${SERVICE_NAME}", BindingServiceTotalServicePoint, 0),
-                new MeasurementPoint("lifecycleTime_service_total_${SERVICE_NAME}", LifecycleTimeServiceTotalServicePoint, 0),
-                new MeasurementPoint("provisionRequest_service_total_${SERVICE_NAME}", ProvisionRequestServiceTotalServicePoint, 0),
-                new MeasurementPoint("provisionRequest_service_success_${SERVICE_NAME}", ProvisionRequestServiceSuccessServicePoint, 0),
-                new MeasurementPoint("provisionRequest_service_fail_${SERVICE_NAME}", ProvisionRequestServiceFailServicePoint, 0),
-                new MeasurementPoint("provisionRequest_plan_total_${PLAN_NAME}", ProvisionRequestPlanTotalPlanPoint, 0),
-                new MeasurementPoint("provisionRequest_plan_success_${PLAN_NAME}", ProvisionRequestPlanSuccessPlanPoint, 0),
-                new MeasurementPoint("provisionRequest_plan_fail_${PLAN_NAME}", ProvisionRequestPlanFailPlanPoint, 0),
-                new MeasurementPoint("provisionedInstances_service_total_${SERVICE_NAME}", ProvisionedInstancesServiceTotalServicePoint, 0),
-                new MeasurementPoint("provisionedInstances_service_success_${SERVICE_NAME}", ProvisionedInstancesServiceSuccessServicePoint, 0),
-                new MeasurementPoint("provisionedInstances_service_fail_${SERVICE_NAME}", ProvisionedInstancesServiceFailServicePoint, 0),
-                new MeasurementPoint("provisionedInstances_plan_total_${PLAN_NAME}", ProvisionedInstancesPlanTotalPlanPoint, 0),
-                new MeasurementPoint("provisionedInstances_plan_success_${PLAN_NAME}", ProvisionedInstancesPlanSuccessPlanPoint, 0),
-                new MeasurementPoint("provisionedInstances_plan_fail_${PLAN_NAME}", ProvisionedInstancesPlanFailPlanPoint, 0)
-        ]
-
-        and:
-        metrics.each { metric ->
-            def query = new Query("select value from ${metric.name}", DB_NAME)
-            def queryResult = influxDB.query(query)
-            def result = influxDBResultMapper.toPOJO(queryResult, metric.measurementPointClass)
-            assert (result.size() == metric.expectedReturnValue)
-        }
-
-        when:
-        serviceBrokerClientExtended.createOrUpdateServiceDefinition(Resource.readTestFileContent("/service-data/service1.json"))
-        Thread.sleep(WAIT_FOR_SERVICE_DEFINITION_TO_BE_DONE_IN_MILLISECONDS)
-        Thread.sleep(TIME_TO_WAIT_FOR_WRITING_TO_INFLUXDB_TO_OCCUR_IN_MILLISECONDS)
-
-        then:
-        metrics.each { metric ->
-            def query = new Query("select value from ${metric.name}", DB_NAME)
-            def queryResult = influxDB.query(query)
-            List<GeneralMeasurement> result = influxDBResultMapper.toPOJO(queryResult, metric.measurementPointClass)
-            assert (result.get(result.size() - 1).value == metric.expectedReturnValue)
-        }
-
-        cleanup:
-        cfServiceRepository.delete(cfServiceRepository.findByGuid(SERVICE_GUID))
-        planRepository.delete(planRepository.findByGuid(PLAN_GUID))
     }
 
     def "update value for total bindings and bindings per service upon binding a service"() {
         given:
-        def metrics = [
-                new MeasurementPoint("bindingRequest_service_fail_${SERVICE_NAME}", BindingRequestServiceFailServicePoint, 0),
-                new MeasurementPoint("bindingRequest_service_success_${SERVICE_NAME}", BindingRequestServiceSuccessServicePoint, 1),
-                new MeasurementPoint("bindingRequest_service_total_${SERVICE_NAME}", BindingRequestServiceTotalServicePoint, 1),
-                new MeasurementPoint("binding_service_total_${SERVICE_NAME}", BindingServiceTotalServicePoint, 1),
-                new MeasurementPoint("binding_total_total", BindingTotalTotalPoint, 1),
+
+        def metricsRetrievableFromDB = [
+                new MeasurementPoint(name: BindingMetricsService.BINDING_SERVICE_KEY, expectedReturnValue: 2),
+                new MeasurementPoint(name: BindingMetricsService.NEW_SERVICE_BINDINGS_KEY, expectedReturnValue: 2)
         ]
 
         when:
-        serviceBrokerClientExtended.createOrUpdateServiceDefinition(Resource.readTestFileContent("/service-data/service1.json"))
-        Thread.sleep(WAIT_FOR_SERVICE_DEFINITION_TO_BE_DONE_IN_MILLISECONDS)
-
-        and:
         def serviceInstanceGuid = UUID.randomUUID().toString()
         serviceBrokerClientExtended.createServiceInstance(new CreateServiceInstanceRequest(SERVICE_GUID, PLAN_GUID, null, null, null).withServiceInstanceId(serviceInstanceGuid).withAsyncAccepted(true))
         def serviceBindingId = UUID.randomUUID().toString()
         serviceBrokerClientExtended.createServiceInstanceBinding(new CreateServiceInstanceBindingRequest(SERVICE_GUID, PLAN_GUID, null, null).withServiceInstanceId(serviceInstanceGuid).withBindingId(serviceBindingId))
+        def serviceBindingId2 = UUID.randomUUID().toString()
+        serviceBrokerClientExtended.createServiceInstanceBinding(new CreateServiceInstanceBindingRequest(SERVICE_GUID, PLAN_GUID, null, null).withServiceInstanceId(serviceInstanceGuid).withBindingId(serviceBindingId2))
         Thread.sleep(TIME_TO_WAIT_FOR_WRITING_TO_INFLUXDB_TO_OCCUR_IN_MILLISECONDS)
 
         then:
-        metrics.each { metric ->
-            def query = new Query("select value from ${metric.name}", DB_NAME)
+        metricsRetrievableFromDB.each { metric ->
+            def query = new Query("select value from ${metric.name} WHERE time >= time() - 1m GROUP BY time(5m)", serviceBrokerMetricsConfig.db)
             def queryResult = influxDB.query(query)
             List<GeneralMeasurement> result = influxDBResultMapper.toPOJO(queryResult, metric.measurementPointClass)
-            assert (result.get(result.size() - 1).value == metric.expectedReturnValue)
+
+            Double value = result.get(result.size() - 1).value
+            println("#### Value is ${value}")
+            assert (value >= metric.expectedReturnValue)
         }
 
         cleanup:
         Thread.sleep(10000)
         serviceBindingRepository.delete(serviceBindingRepository.findByGuid(serviceBindingId))
+        serviceBindingRepository.delete(serviceBindingRepository.findByGuid(serviceBindingId2))
         serviceInstanceRepository.delete(serviceInstanceRepository.findByGuid(serviceInstanceGuid))
-        cfServiceRepository.delete(cfServiceRepository.findByGuid(SERVICE_GUID))
-        planRepository.delete(planRepository.findByGuid(PLAN_GUID))
     }
 
     def "update value for total lifecycle time per service upon deprovisioning a service instance"() {
@@ -210,10 +148,6 @@ class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
         ]
 
         when:
-        serviceBrokerClientExtended.createOrUpdateServiceDefinition(Resource.readTestFileContent("/service-data/service1.json"))
-        Thread.sleep(WAIT_FOR_SERVICE_DEFINITION_TO_BE_DONE_IN_MILLISECONDS)
-
-        and:
         def serviceInstanceGuid = UUID.randomUUID().toString()
         serviceBrokerClientExtended.createServiceInstance(new CreateServiceInstanceRequest(SERVICE_GUID, PLAN_GUID, null, null, null).withServiceInstanceId(serviceInstanceGuid).withAsyncAccepted(true))
         Thread.sleep(LIFECYCLE_TIME_IN_MILLISECONDS)
@@ -224,7 +158,7 @@ class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
 
         then:
         metrics.each { metric ->
-            def query = new Query("select value from ${metric.name}", DB_NAME)
+            def query = new Query("select value from ${metric.name}", serviceBrokerMetricsConfig.db)
             def queryResult = influxDB.query(query)
             List<GeneralMeasurement> result = influxDBResultMapper.toPOJO(queryResult, metric.measurementPointClass)
             assertEquals(result.get(result.size() - 1).value, metric.expectedReturnValue, 1000)
@@ -232,8 +166,6 @@ class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
 
         cleanup:
         serviceInstanceRepository.delete(serviceInstanceRepository.findByGuid(serviceInstanceGuid))
-        cfServiceRepository.delete(cfServiceRepository.findByGuid(SERVICE_GUID))
-        planRepository.delete(planRepository.findByGuid(PLAN_GUID))
     }
 
     def "update value for provision requests and provisioned instances including values for services and plans upon provisioning a service instance"() {
@@ -265,17 +197,13 @@ class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
         ]
 
         when:
-        serviceBrokerClientExtended.createOrUpdateServiceDefinition(Resource.readTestFileContent("/service-data/service1.json"))
-        Thread.sleep(WAIT_FOR_SERVICE_DEFINITION_TO_BE_DONE_IN_MILLISECONDS)
-
-        and:
         def serviceInstanceGuid = UUID.randomUUID().toString()
         serviceBrokerClientExtended.createServiceInstance(new CreateServiceInstanceRequest(SERVICE_GUID, PLAN_GUID, null, null, null).withServiceInstanceId(serviceInstanceGuid).withAsyncAccepted(true))
         Thread.sleep(TIME_TO_WAIT_FOR_WRITING_TO_INFLUXDB_TO_OCCUR_IN_MILLISECONDS)
 
         then:
         metrics.each { metric ->
-            def query = new Query("select value from ${metric.name}", DB_NAME)
+            def query = new Query("select value from ${metric.name}", serviceBrokerMetricsConfig.db)
             def queryResult = influxDB.query(query)
             List<GeneralMeasurement> result = influxDBResultMapper.toPOJO(queryResult, metric.measurementPointClass)
             assert (result.get(result.size() - 1).value == metric.expectedReturnValue)
@@ -284,8 +212,6 @@ class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
         cleanup:
         Thread.sleep(10000)
         serviceInstanceRepository.delete(serviceInstanceRepository.findByGuid(serviceInstanceGuid))
-        cfServiceRepository.delete(cfServiceRepository.findByGuid(SERVICE_GUID))
-        planRepository.delete(planRepository.findByGuid(PLAN_GUID))
     }
 
     def "ignore deleted service instances in metrics regarding provisioned instances but include in provision requests"() {
@@ -317,10 +243,6 @@ class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
         ]
 
         when:
-        serviceBrokerClientExtended.createOrUpdateServiceDefinition(Resource.readTestFileContent("/service-data/service1.json"))
-        Thread.sleep(WAIT_FOR_SERVICE_DEFINITION_TO_BE_DONE_IN_MILLISECONDS)
-
-        and:
         def serviceInstanceGuid = UUID.randomUUID().toString()
         serviceBrokerClientExtended.createServiceInstance(new CreateServiceInstanceRequest(SERVICE_GUID, PLAN_GUID, null, null, null).withServiceInstanceId(serviceInstanceGuid).withAsyncAccepted(true))
 
@@ -330,16 +252,13 @@ class ServiceBrokerMetricsServiceFunctionalSpec extends BaseFunctionalSpec {
 
         then:
         metrics.each { metric ->
-            def query = new Query("select value from ${metric.name}", DB_NAME)
+            def query = new Query("select value from ${metric.name}", serviceBrokerMetricsConfig.db)
             def queryResult = influxDB.query(query)
-            List<GeneralMeasurement> result = influxDBResultMapper.toPOJO(queryResult, metric.measurementPointClass)
+            def result = influxDBResultMapper.toPOJO(queryResult, metric.measurementPointClass)
             assert (result.get(result.size() - 1).value == metric.expectedReturnValue)
         }
 
         cleanup:
         serviceInstanceRepository.delete(serviceInstanceRepository.findByGuid(serviceInstanceGuid))
-        cfServiceRepository.delete(cfServiceRepository.findByGuid(SERVICE_GUID))
-        planRepository.delete(planRepository.findByGuid(PLAN_GUID))
     }
-
 }